Feat: 닉네임 정규식 추가 및 예외 처리 (#4)
